rename: components 구조 변경
* DocumentEditPage와 Sidebar로 나누기
